+++ .buzzy/package.yaml
@@ -1,4 +1,4 @@
-name: ipset
+name: libcorkipset
 build_dependencies:
   - pkg-config
   - check >= 0.9.4
