+++ README.markdown
@@ -2,7 +2,7 @@
 
 [![Build Status](https://img.shields.io/travis/redjack/ipset/develop.svg)](https://travis-ci.org/redjack/ipset)
 
-The ipset library provides C data types for storing sets of IP
+The libcorkipset library provides C data types for storing sets of IP
 addresses, and maps of IP addresses to integers.  It supports both
 IPv4 and IPv6 addresses.  It's implemented using [Binary Decision
 Diagrams](http://en.wikipedia.org/wiki/Binary_decision_diagram)
