+++ src/libcorkipset/general.c
@@ -1,0 +1,20 @@
+/* -*- coding: utf-8 -*-
+ * ----------------------------------------------------------------------
+ * Copyright © 2009, libcorkipset authors
+ * All rights reserved.
+ *
+ * Please see the COPYING file in this distribution for license details.
+ * ----------------------------------------------------------------------
+ */
+
+#include <libcork/core.h>
+
+#include "libcork/ipset.h"
+#include "libcork/ipset/nodes.h"
+
+
+int
+ipset_init_library()
+{
+    return 0;
+}
