+++ tests/test-assignment.c
@@ -13,8 +13,8 @@
 #include <check.h>
 #include <libcork/core.h>
 
-#include "ipset/bdd/nodes.h"
-#include "ipset/bits.h"
+#include "libcork/ipset/bits.h"
+#include "libcork/ipset/nodes.h"
 
 
 /*-----------------------------------------------------------------------
