+++ tests/test-bdd.c
@@ -13,8 +13,8 @@
 #include <check.h>
 #include <libcork/core.h>
 
-#include "ipset/bdd/nodes.h"
-#include "ipset/bits.h"
+#include "libcork/ipset/bits.h"
+#include "libcork/ipset/nodes.h"
 
 
 #define DESCRIBE_TEST  fprintf(stderr, "---\n%s\n", __func__)
